+++ ukis_pysat/members.py
@@ -11,6 +11,7 @@ class Platform(Enum):
 
 
 class Datahub(Enum):
-    File = auto()
-    EarthExplorer = auto()
-    Scihub = auto()
+    STAC_local = "STAC Catalog/Collection"
+    EarthExplorer = "EarthExplorer"
+    Scihub = "Scihub"
+    STAC_API = "STAC API"